src/components/Stepper/stories/hooks/useHorizontalLinearStepper.ts
import { useState, ReactNode } from 'react';

export const useHorizontalLinearStepper = () => {
  const [activeStep, setActiveStep] = useState(0);
  const [skipped, setSkipped] = useState(new Set<number>());
  const steps = [
    'Select campaign settings',
    'Create an ad group',
    'Create an ad',
  ];
  const stepProps: { completed?: boolean } = {};
  const labelProps: {
    optional?: ReactNode;
  } = {};

  const isStepOptional = (step: number) => {
    return step === 1;
  };

  const isStepSkipped = (step: number) => {
    return skipped.has(step);
  };

  const handleNext = () => {
    let newSkipped = skipped;
    if (isStepSkipped(activeStep)) {
      newSkipped = new Set(newSkipped.values());
      newSkipped.delete(activeStep);
    }

    setActiveStep((prevActiveStep) => prevActiveStep + 1);
    setSkipped(newSkipped);
  };

  const handleBack = () => {
    setActiveStep((prevActiveStep) => prevActiveStep - 1);
  };

  const handleSkip = () => {
    if (!isStepOptional(activeStep)) {
      throw new Error("You can't skip a step that isn't optional.");
    }

    setActiveStep((prevActiveStep) => prevActiveStep + 1);
    setSkipped((prevSkipped) => {
      const newSkipped = new Set(prevSkipped.values());
      newSkipped.add(activeStep);
      return newSkipped;
    });
  };

  const handleReset = () => {
    setActiveStep(0);
  };
  return {
    data: { steps, activeStep, labelProps, stepProps },
    actions: {
      isStepOptional,
      isStepSkipped,
      handleReset,
      handleBack,
      handleNext,
      handleSkip,
    },
  };
};
